+++ CHANGELOG.md
@@ -62,7 +62,7 @@
 * Changes section titles in Instance form. UIIN-72
 * Migrate to hierarchical location end-point. Fixes UIIN-127.
 * Handle "metaData" field case insensitively for compatibility. Fixes UIIN-159.
-* Only show item-availabilty for open loans. Refs UIIN-110.
+* Only show item-availability for open loans. Refs UIIN-110.
 * Add location-units and locations to Okapi interfaces. Fixes UIIN-171.
 * Correct use of i18n in validation function and props. Fixes UIIN-172.
 * Confirmation when assigning inactive location. Fixes UIIN-121.
@@ -72,6 +72,7 @@
 * Enable opening item view and holdings view in new tab. UIIN-147, UIIN-148.
 * Use a single handler for managing row-level links in MCL. Refs UIIN-122.
 * Replace shelf location references with locations. Fixes UIIN-182.
+* Added `code` and `source` fields to several settings pages. Fixes UIIN-149, UIIN-150, UIIN-151, UIIN-152,
 
 ## [1.0.0](https://github.com/folio-org/ui-instances/tree/v1.0.0) (2017-09-08)
 [Full Changelog](https://github.com/folio-org/ui-items/compare/v0.0.1...v1.0.0)